Made gift card grid items clickable
Defined a rule to make .selected-offer-gift-card have a blue border. Also
added a rule to give a border with the same color as the page background color
to non-active gift cards to prevent reflow when a grid item is clicked.

Added a temporary useState in the VerticalOffers component to record the
activeOfferId. This will be replaced when we change the redux state shape
to contain the active offer.