Add a key to the checkout slice of the redux state
Added checkoutOffer key to the checkout slice of the redux state. Removed
local state from the VerticalOffers component; the logic to set the
current checkout offer was moved up to the DisplayOffers component,
which now passes the activeOfferId to VerticalOffers.

Commented out a type in offers-slice.ts to prevent a typescript related
lint error.